Got rid of a bunch of warnings from mcs 1.1.6. Guarded finalizers and     FinalizerAssert with HasShutdownStarted. Updated INSTALL.MONO and     CHANGES file.

+++ cs/src/Ice/CommunicatorI.cs
@@ -296,17 +296,13 @@ namespace Ice
 #if DEBUG
 	~CommunicatorI()
 	{
-            /**
-              * We cannot invoke methods on other objects in a destructor.
-              *
 	    lock(this)
 	    {
-		if(!_destroyed)
+		if(!_destroyed && !Environment.HasShutDownStarted)
 		{
 		    _instance.logger().warning("Ice::Communicator::destroy() has not been called");
 		}
 	    }
-              **/
 	}
 #endif
